// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  ADMIN: 'ADMIN',
  DOCTOR: 'DOCTOR',
  PATIENT: 'PATIENT'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  BLOCKED: 'BLOCKED',
  DELETED: 'DELETED',
  ACTIVE: 'ACTIVE'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const Gender = {
  MALE: 'MALE',
  FEMALE: 'FEMALE'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const AppoinmentStatus = {
  SCHEDULED: 'SCHEDULED',
  INPROGRESS: 'INPROGRESS',
  COMPLETED: 'COMPLETED'
} as const

export type AppoinmentStatus = (typeof AppoinmentStatus)[keyof typeof AppoinmentStatus]


export const PaymentStatus = {
  PAID: 'PAID',
  UNPAID: 'UNPAID'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]


export const BloodGroup = {
  A_POSITIVE: 'A_POSITIVE',
  A_NEGATIVE: 'A_NEGATIVE',
  B_POSITIVE: 'B_POSITIVE',
  B_NEGATIVE: 'B_NEGATIVE',
  AB_POSITIVE: 'AB_POSITIVE',
  AB_NEGATIVE: 'AB_NEGATIVE',
  O_POSITIVE: 'O_POSITIVE',
  O_NEGATIVE: 'O_NEGATIVE'
} as const

export type BloodGroup = (typeof BloodGroup)[keyof typeof BloodGroup]


export const MaritalStaus = {
  MARRIED: 'MARRIED',
  UNMARRIED: 'UNMARRIED',
  DIVORCED: 'DIVORCED',
  WIDOWED: 'WIDOWED'
} as const

export type MaritalStaus = (typeof MaritalStaus)[keyof typeof MaritalStaus]